Split schema and main tests
